refactor(llm): move anthropic tool_search_passthrough into model_constructors
