+++ front/lib/model_constructors/sdk/anthropic_ai/converters/input/tool_search_passthrough.ts
