+++ front/lib/model_constructors/sdk/anthropic_ai/tool_search_dangling_server_tool_use.test.ts
@@ -6,7 +6,7 @@ import type { MessageParam } from "@anthropic-ai/sdk/resources/messages/messages
 import {
   parseAnthropicToolSearchBlock,
   stripUnreplayableToolSearchBlocks,
-} from "@app/lib/api/llm/clients/anthropic/utils/tool_search_passthrough";
+} from "@app/lib/model_constructors/sdk/anthropic_ai/converters/input/tool_search_passthrough";
 import type { MessageBlockConverters } from "@app/lib/model_constructors/sdk/anthropic_ai/converters/input/utils";
 import {
   assistantProviderPassthroughMessageToBlocks,
